requirements.txt
# --------- core ML dependencies --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- experiment management --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
wandb>=0.15.0  # Specific version for WandB as primary logger
hydra-optuna-sweeper==1.2.0  # For local hyperparameter sweeps

# --------- code quality --------- #
black>=23.0.0  # Code formatting
ruff>=0.1.0    # Fast linter
pytest>=7.0.0  # Testing

# --------- project management --------- #
invoke>=2.0.0  # Task management
rich>=13.0.0   # Terminal output formatting
python-dotenv>=1.0.0  # Environment variable management

# --------- type checking --------- #
mypy>=1.0.0    # Static type checking
types-PyYAML   # Type stubs for PyYAML
